requirements.txt
# SPECTRE-SIM Python Dependencies
# Install with: pip install -r requirements.txt
# Tested on Python 3.10, 3.11, 3.12

# Core scientific stack
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Progress bar for Monte Carlo sweeps
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# YAML configuration
PyYAML>=6.0
